Fix for session clashes
Sessions were clashing when a user was managing the same app for
multiple stores in the same browser.

Both stores have the same session due to them both using the app
domain in the same browser.

This was leading to store A overwriting store B or vice-versa.

The solution is to check the referer header to see exactly which
store we're dealing with.

Because the referer header is spoofable we make sure that there is
an HMAC signature present in the referer.

If there is a shop domain with a valid HMAC signature in the referer
header then we overwrite the shopDomain variable to make sure the
correct store is loaded.